Major refactoring: - Splitting P360 from FINT and NOARK - Cleanup
