+++ src/main/java/no/fint/ra/data/exception/CreateCaseException.java
@@ -1,0 +1,7 @@
+package no.fint.ra.data.exception;
+
+public class CreateCaseException extends RuntimeException {
+    public CreateCaseException(String message) {
+        super(message);
+    }
+}
