--- src/main/java/no/fint/ra/data/exception/CreateTilskuddFartoyException.java
@@ -1,7 +1,0 @@
-package no.fint.ra.data.exception;
-
-public class CreateTilskuddFartoyException extends RuntimeException {
-    public CreateTilskuddFartoyException(String message) {
-        super(message);
-    }
-}
