--- src/main/java/no/fint/ra/data/p360/P360CaseFactory.java
@@ -1,95 +1,0 @@
-package no.fint.ra.data.p360;
-
-import no.fint.arkiv.p360.caze.*;
-import no.fint.model.resource.kultur.kulturminnevern.TilskuddFartoyResource;
-import no.fint.ra.KulturminneProps;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import javax.annotation.PostConstruct;
-import javax.xml.bind.JAXBElement;
-import java.util.Arrays;
-import java.util.List;
-
-@Component
-public class P360CaseFactory {
-
-    @Autowired
-    private KulturminneProps kulturminneProps;
-
-    private ObjectFactory objectFactory;
-
-
-    @PostConstruct
-    private void init() {
-        objectFactory = new ObjectFactory();
-    }
-
-    public CreateCaseParameter createTilskuddFartoy(TilskuddFartoyResource tilskuddFartoy) {
-        CreateCaseParameter createCaseParameter = new CreateCaseParameter();
-
-        createCaseParameter.setTitle(objectFactory.createCaseParameterBaseTitle(tilskuddFartoy.getTittel()));
-        createCaseParameter.setStatus(objectFactory.createCaseParameterBaseStatus("B"));
-        createCaseParameter.setFiledOnPaper(objectFactory.createCaseParameterBaseFiledOnPaper(false));
-
-        createCaseParameter.setKeywords(objectFactory.createCaseParameterBaseKeywords(getKeywords()));
-
-        // Set default to NOARK Sak
-        createCaseParameter.setCaseType(objectFactory.createCreateCaseParameterCaseType("recno:2"));
-
-        // TODO: 2019-04-30 Denne bør vel egentlig komme fra journalEnhet
-        createCaseParameter.setResponsibleEnterpriseRecno(objectFactory.createCaseParameterBaseResponsibleEnterpriseRecno(Integer.valueOf(kulturminneProps.getResponsibleUnit())));
-        createCaseParameter.setSubArchive(objectFactory.createCaseParameterBaseSubArchive(kulturminneProps.getSubArchive()));
-
-        ExternalIdParameter externalIdParameter = objectFactory.createExternalIdParameter();
-        externalIdParameter.setId(objectFactory.createExternalIdParameterId(tilskuddFartoy.getSoknadsnummer().getIdentifikatorverdi()));
-        externalIdParameter.setType(objectFactory.createExternalIdParameterType("Digisak"));
-
-        createCaseParameter.setExternalId(objectFactory.createCaseParameterBaseExternalId(externalIdParameter));
-        ArrayOfClassCodeParameter arrayOfClassCodeParameter = objectFactory.createArrayOfClassCodeParameter();
-        ClassCodeParameter classCodeParameter = objectFactory.createClassCodeParameter();
-
-        classCodeParameter.setSort(1);
-        classCodeParameter.setIsManualText(Boolean.FALSE);
-        classCodeParameter.setArchiveCode(objectFactory.createClassCodeParameterArchiveCode(tilskuddFartoy.getFartoyNavn()));
-        classCodeParameter.setArchiveType(objectFactory.createClassCodeParameterArchiveType("Fartøy"));
-        arrayOfClassCodeParameter.getClassCodeParameter().add(classCodeParameter);
-
-        createCaseParameter.setArchiveCodes(objectFactory.createCaseParameterBaseArchiveCodes(arrayOfClassCodeParameter));
-
-        /*
-        ArrayOfRemark arrayOfRemark = objectFactory.createArrayOfRemark();
-        Remark remark = new Remark();
-        remark.setRemarkType(objectFactory.createString("SI"));
-        remark.setContent(objectFactory.createString("test"));
-        arrayOfRemark.getRemark().add(remark);
-        createCaseParameter.setRemarks(objectFactory.createArrayOfRemark(arrayOfRemark));
-        */
-        /*
-        createCaseParameter.setResponsiblePersonIdNumber(
-                objectFactory.createCaseParameterBaseResponsiblePersonIdNumber(
-                        tilskuddFartoy.getSaksansvarlig().get(0).getHref()
-                )
-        );
-        */
-
-
-
-
-        return createCaseParameter;
-
-    }
-
-    private ArrayOfstring getKeywords() {
-
-        ArrayOfstring keywordArray = objectFactory.createArrayOfstring();
-        List<String> keywords = Arrays.asList(kulturminneProps.getKeywords());
-
-        keywords.forEach(keywordArray.getString()::add);
-
-        return keywordArray;
-    }
-
-
-
-}
