+++ src/main/java/no/fint/ra/data/utilities/Constants.java
@@ -1,0 +1,8 @@
+package no.fint.ra.data.utilities;
+
+public enum Constants {
+    ;
+
+    public static final String EXTERNAL_ID_TYPE = "Digisak";
+    public static final String CASE_TYPE_NOARK = "recno:2";
+}
